+++ src/containers/week/index.js
@@ -1,14 +1,22 @@
 import React from 'react';
 import { connect } from 'react-redux';
-import WeekViewComponent, { NDaysPicker } from '../../components/week';
+import { sortEventsByDate } from 'react-native-week-view';
+import WeekViewComponent from '../../components/week';
+import NDaysPicker from '../../components/week/nDaysPicker';
+import { Memoizer } from '../../shared/utils';
 import { workSessionsSelector } from '../../redux/selectors';
 import { getToday } from '../../shared/dates';
 import { HeaderActions } from '../../shared/UI/headers';
 import i18n from '../../shared/i18n';
+import { getLightColor } from '../../shared/styles';
 
 export class WeekView extends React.Component {
   static navigationOptions({ navigation }) {
     const actions = [
+      {
+        icon: 'today',
+        handlePress: navigation.getParam('handlePressGoToToday'),
+      },
       {
         icon: 'columns',
         handlePress: navigation.getParam('handlePressNDaysMenu'),
@@ -28,52 +36,41 @@ export class WeekView extends React.Component {
       nDays: 7,
       nDaysMenuVisible: false,
       selectedDate: getToday(),
+      workSessionsByDate: {},
+      isProcessing: true,
     };
+    // NOTE: the selectedDate value won't be updated with the current selectedDate,
+    // as the event handlers from react-native-week-view are not connected.
+    // If they were connected (onSwipePrev/Next), on each call, a setState() should
+    // be called, which would trigger an un-necessary re-render.
 
     this.handlePressNDaysMenu = this.handlePressNDaysMenu.bind(this);
     this.handleChangeNDays = this.handleChangeNDays.bind(this);
+    this.handlePressGoToToday = this.handlePressGoToToday.bind(this);
     this.closeModal = this.closeModal.bind(this);
-    this.forceReloadIfChanged = this.forceReloadIfChanged.bind(this);
+    this.processWorkSessions = this.processWorkSessions.bind(this);
 
     this.props.navigation.setParams({
       handlePressNDaysMenu: this.handlePressNDaysMenu,
+      handlePressGoToToday: this.handlePressGoToToday,
     });
 
-    this.workSessionsChanged = false;
+    this.memoizer = Memoizer();
   }
 
   componentDidMount() {
-    this.willFocusListener = this.props.navigation.addListener(
-      'willFocus',
-      this.forceReloadIfChanged,
+    this.didFocusListener = this.props.navigation.addListener(
+      'didFocus',
+      () => this.processWorkSessions(),
     );
   }
 
   shouldComponentUpdate(nextProps) {
-    // NOTE/HACK: the (forceReloadIfChanged + workSessionsChanged +
-    // shouldComponentUpdate) logic is done to avoid heavy renders when the tab
-    // is not focused. Rendering of children is heavy. When the user is at
-    // another tab and modifies workSessions (e.g. play/pause/stop, delete
-    // workSession) a re-render of said children will be triggered, causing a
-    // lot of latency. To fix this, shouldComponentUpdate indicates to update
-    // only when the tab is focused. However, this implies that when the tab is
-    // focused again, the workSessions prop will be outdated (it wasn't updated
-    // when shouldComponentUpdate returned false!). Hence, a forceUpdate is
-    // called on willFocus, but only if the workSessions prop actually changed.
-    const { workSessions } = this.props;
-    this.workSessionsChanged = workSessions !== nextProps.workSessions;
     return nextProps.navigation.isFocused();
   }
 
   componentWillUnmount() {
-    if (this.willFocusListener) this.willFocusListener.remove();
-  }
-
-  forceReloadIfChanged() {
-    if (this.workSessionsChanged) {
-      this.workSessionsChanged = false;
-      this.forceUpdate();
-    }
+    if (this.didFocusListener) this.didFocusListener.remove();
   }
 
   handlePressNDaysMenu() {
@@ -85,6 +82,13 @@ export class WeekView extends React.Component {
   handleChangeNDays(nDays) {
     this.setState({
       nDays,
+      nDaysMenuVisible: false,
+    });
+  }
+
+  handlePressGoToToday() {
+    this.setState({
+      selectedDate: getToday(),
     });
   }
 
@@ -94,17 +98,51 @@ export class WeekView extends React.Component {
     });
   }
 
-  render() {
+  processWorkSessions() {
     const { workSessions } = this.props;
+
+    if (!this.memoizer.hasChanged(workSessions)) {
+      this.setState({
+        isProcessing: false,
+      });
+      return;
+    }
+
+    this.setState({
+      isProcessing: true,
+    });
+
+    requestAnimationFrame(() => {
+      const events = workSessions.map((workSession) => {
+        const { subject } = workSession;
+        const { category } = subject;
+        const color = category ? category.color : 'gray'; // HACK: default hardcoded
+        return {
+          id: workSession.id,
+          description: subject.name,
+          startDate: workSession.getLocalStartDate(),
+          endDate: workSession.getLocalEndDate(),
+          color: getLightColor(color),
+        };
+      });
+      this.setState({
+        workSessionsByDate: sortEventsByDate(events),
+        isProcessing: false,
+      });
+    });
+  }
+
+  render() {
     const {
-      nDays, nDaysMenuVisible, selectedDate,
+      nDays, nDaysMenuVisible, selectedDate, workSessionsByDate, isProcessing,
     } = this.state;
 
     return (
       <WeekViewComponent
-        workSessions={workSessions}
+        workSessions={workSessionsByDate}
         nDays={nDays}
         selectedDate={selectedDate}
+        isProcessing={isProcessing}
       >
         <NDaysPicker
           isVisible={nDaysMenuVisible}