+++ src/shared/i18n/calendars/index.js
@@ -3,3 +3,9 @@ import es from './es';
 
 LocaleConfig.locales.es = es;
 LocaleConfig.locales.en = LocaleConfig.locales[''];
+
+const setRNCalendarsLocale = (locale) => {
+  LocaleConfig.defaultLocale = locale;
+};
+
+export default setRNCalendarsLocale;