+++ docs/CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+* WeekView loading spinner now covers all cases
+* Add i18n for WeekView
+
 
 
 ## 1.12.0 - 2020-05-21